+++ src/main/views/features/mediation/i-dont-want-free-mediation.njk
@@ -3,8 +3,8 @@
 {% from "govuk/components/button/macro.njk" import govukButton %}
 {% from "govuk/components/radios/macro.njk" import govukRadios %}
 {% from "govuk/components/textarea/macro.njk" import govukTextarea %}
-{% from "../../../macro/contact-us-for-help.njk" import contactUsForHelp %}
-{% from "../../../macro/csrf.njk" import csrfProtection %}
+{% from "../../macro/contact-us-for-help.njk" import contactUsForHelp %}
+{% from "../../macro/csrf.njk" import csrfProtection %}
 
 {% block content %}
   <div class="govuk-width-container">
